remove HttpHelper and move logic to ZuliaRESTClient
